+++ packages/cli-kit/src/public/node/plugins/tunnel.ts
@@ -1,8 +1,20 @@
-import {err, Result} from '../result.js'
 import {ExtendableError} from '../error.js'
 import {FanoutHookFunction, PluginReturnsForHook} from '../plugins.js'
+import {err, Result} from '../result.js'
 
 export type TunnelErrorType = 'invalid-provider' | 'tunnel-already-running' | 'wrong-credentials' | 'unknown'
+export interface TunnelClient {
+  getTunnelStatus: () => TunnelStatusType
+  stopTunnel: () => void
+  provider: string
+  port: number
+}
+export type TunnelStatusType =
+  | {status: 'not-started'}
+  | {status: 'starting'}
+  | {status: 'connected'; url: string}
+  | {status: 'error'; message: string}
+
 export class TunnelError extends ExtendableError {
   type: TunnelErrorType
   constructor(type: TunnelErrorType, message?: string) {
@@ -20,7 +32,7 @@ export interface HookReturnPerTunnelPlugin {
   tunnel_start: {
     options: {port: number; provider: string}
     pluginReturns: {
-      [pluginName: string]: Result<{url: string}, TunnelError>
+      [key: string]: Result<TunnelClient, TunnelError>
     }
   }
   tunnel_provider: {
@@ -39,6 +51,7 @@ export type TunnelStartAction = (port: number) => Promise<TunnelStartReturn>
 export const defineProvider = (input: {name: string}): TunnelProviderFunction => {
   return async () => input
 }
+
 export const startTunnel = (options: {provider: string; action: TunnelStartAction}): TunnelStartFunction => {
   return async (inputs: {provider: string; port: number}): Promise<TunnelStartReturn> => {
     if (inputs.provider !== options.provider) return err(new TunnelError('invalid-provider'))